cleaned up old files

+++ combined/db_codes/procedure.sql
@@ -70,7 +70,7 @@ where evia.`Unnamed: 0` NOT IN
 drop procedure if exists searchMaxThreeSelectedEvidence;
 DELIMITER $$
 create procedure searchMaxThreeSelectedEvidence(IN drugName VARCHAR(255), IN evi1 VARCHAR(255), IN evi2 VARCHAR(255), 
-IN evi3 VARCHAR(255), IN threshold int)
+IN evi3 VARCHAR(255), IN threshold FLOAT)
 begin
     DECLARE done int default false;
     DECLARE table_name CHAR(255);
